backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
pillow==10.4.0
requests==2.32.3
httpx==0.27.2
websockets==13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1   # passlib 1.7.4 no soporta bcrypt >=4.1
aiofiles==24.1.0
python-dateutil==2.9.0
apscheduler==3.10.4
reportlab==4.2.5
slowapi==0.1.9
email-validator==2.2.0

# Tests (sólo dev)
pytest==8.3.3
pytest-asyncio==0.24.0
